requirements.txt
numpy==1.21.0
deepdanbooru-onnx>=0.0.8
fastapi>=0.89.1
pillow>=9.4.0
uvicorn[standard]>=0.20.0
python-multipart>=0.0.5
setuptools>=57.0.0
wheel>=0.36.2
Jinja2>=3.0.1
tensorflow>=2.0.0
